requirements/dev.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
comm==0.1.2
constantly==15.1.0
cryptography==39.0.0
cssselect==1.2.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
fqdn==1.5.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipdb==0.13.11
ipykernel==6.19.4
ipython==8.8.0
ipython-genutils==0.2.0
isoduration==20.11.0
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
kaitaistruct==0.10
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
outcome==1.2.0
packaging==22.0
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
path==16.6.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
Protego==0.2.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.14.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
PyYAML==6.0
python-arango==6.0.0
pyzmq==24.0.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
selenium==4.7.2
selenium-wire==5.1.0
Send2Trash==1.8.0
scrapy-crawlera==1.3.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tldextract==3.4.0
tornado==6.2
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
w3lib==2.1.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
wsproto==1.2.0
zope.interface==5.5.2
zstandard==0.19.0
